Item Inventory working
Item Inventory working. It adds items correctly and shows ingame. Add a button to the inventory slots so they can work with touch inputs in mobile.
Added new controls: q and e for InventoryScroll in PlayerController (Script)
Added debug controls: r and y for adding or removing a default item in the inventory selected slot and i and o for adding and removing inventory slots (changing maxItems) in Inventory (Script)
Added a new Asset to use it as a Default sprite of a bubble and its respective Item (Create > Item in the Assets/Items folder)